\name{data_Bliese}
\alias{data_Bliese}
\docType{data}
\title{data_Bliese}
\description{A data frame with 750 individual-level observations, nested within 50 groups, on 8 variables.}
\details{A data frame with 750 individual-level observations, nested within 50 groups, on 
	8 variables. This simulated dataset was created by Paul Bliese, and results from WABA 
	analyses of this data were reported by Klein et al. (2000). The dataset is provided as 
	trial data for the waba function.}
\references{
Klein, K., Bliese, B., Kozlowski, S., Dansereau, F., Gavin, M., Griffin, M., Hofmann, D., 
James, L., Yammarino, F., & Bligh, M. (2000). Multilevel analytical techniques: Commonalities, 
differences, and continuing questions. In K. J. Klein & S. W. J. Kozlowski (Eds.), Multilevel 
theory, research and methods in organizations (pp. 512-556). San Francisco, CA: Jossey-Bass. 
}
\examples{summary(data_Bliese) }
