\name{data_Detect_Set_A}
\alias{data_Detect_Set_A}
\docType{data}
\title{data_Detect_Set_A}
\description{A data frame with 40 cases nested within 20 groups.}
\details{This dataset is from Dataset A at www.levelsofanalysis.com, and as described 
	in chapter 5 of the DETECT manual (which can be viewed at that Internet site). 
	The dataset, which is synthetic, contains 40 cases nested within 20 groups. It 
	contains values for DYAD, NEGOT, SATIS, and PERF. It is provided as trial data 
	for the waba and wabamodreg functions.}
\examples{summary(data_Detect_Set_A) }
