\name{data_Detect_Set_A_mra}
\alias{data_Detect_Set_A_mra}
\docType{data}
\title{data_Detect_Set_A_mra}
\description{A data frame with 40 cases nested within 20 groups.}
\details{
This dataset is from Dataset A from www.levelsofanalysis.com, and as described in 
chapter 5 of the DETECT manual (which can be viewed at that Internet site). The data 
set, which is synthetic, contains 40 cases nested within 20 groups. It contains values 
for TASKCLAR, DYAD, NEGOT,and SATIS. It is provided as trial data for the wabamra 
function, which conducts separate WABA analyses for each value of a specified 
"Condition" variable, which in this case is TASKCLAR.}
\examples{summary(data_Detect_Set_A_mra) }
