\name{data_ggdist}
\alias{data_ggdist}
\docType{data}
\title{data_ggdist}
\description{A data frame with 48 observations on 3 variables from 24 dyads.}
\details{A data frame with 48 observations on 3 variables from 24 dyads. It contains
	values for dyad, person, var1, var2, and var3. It is provided as trial data for 
	the ggdist function. }
\examples{summary(data_ggdist) }
