\name{data_ggexch}
\alias{data_ggexch}
\docType{data}
\title{data_ggexch}
\description{A data frame with 48 observations on 3 variables from 24 dyads.}
\details{A data frame with 48 observations on 3 variables from 24 dyads. It contains 
	values for dyad, var1, var2, and var3. It is provided as trial data for the ggexch function. }
\examples{summary(data_ggexch) }
