\name{data_jsp}
\alias{data_jsp}
\docType{data}
\title{data_jsp}
\description{A data frame with observations on 1129 high school students nested within 49 schools.}
\details{A data frame with observations on 1129 high school students nested within 49 schools.
	This the Year 1 data from the Junior School Project dataset that was analyzed by 
	Goldstein (1995). There are values for school, class, sex, pupil, and scores on ravens, 
	english, maths. The dataset is provided as trial data for the waba, wabamra, and 
	wabamodreg functions.}
\references{
Goldstein, H. (1995) Multilevel Statistical Models. New York, NY: Halstead Press. 
}
\examples{summary(data_jsp) }
