\name{waba-package}

\alias{waba-package}

\docType{package}

\title{ Within-And-Between-Analysis For Groups and Dyads }

\description{Levels-of-analysis issues arise whenever individual-level data are collected 
	from more than one person from the same dyad, family, classroom, work group, or other 
	interaction unit. Interdependence among cases from the same units also violates the 
	independence-of-observations assumption that underlies many statistical tests. This 
	package contains functions for addressing these issues via: (1) the 
	within-and-between-analyses (WABA) framework described by Dansereau, Alutto, and 
	Yammarino (1984), and (2) the dyad-level analyses described by Gonzalez and 
	Griffin (1999) and Griffin and Gonzalez (1995). Descriptions of the procedures 
	were provided by Yammarino and Markham (1992), Yammarino (1998), and by O'Connor (2004).
}
\references{
 {  Dansereau, F., Alutto, J., & Yammarino, F. (1984). Theory testing in organizational behavior. 
 	Englewood Cliffs, NJ: Prentice-Hall.}
\cr\cr {Griffin, D., & Gonzalez, R. (1995). Correlational analysis of dyadlevel data in the 
	exchangeable case. Psychological Bulletin, 118, 430-439.}
\cr\cr {Gonzalez, R., & Griffin, D. (1999). The correlational analysis of dyad-level data in 
	the distinguishable case. Personal Relationships, 6, 449-469.}
\cr\cr {O'Connor, B. P. (2004). SPSS and SAS programs for addressing interdependence and basic 
	levels-of-analysis issues in psychological data. Behavior Research Methods, Instrumentation, 
	and Computers, 36, 17-28.}
\cr\cr {Yammarino, F. J. (1998). Multivariate aspects of the varient /WABA approach. Leadership 
	Quarterly, 9, 203-227.}
\cr\cr {Yammarino, F. J., & Markham, S. (1992). On the application of within and between analysis: 
	Are absence and affect really group based. Journal of Applied Psychology, 77, 168-176.}
}

